requirements.txt
requests==2.25.1
Flask==1.1.2
joblib==1.0.1
numpy==1.20.1
Pillow==8.1.0
tensorflow==2.4.1
